+++ NEWS.md
@@ -2,5 +2,16 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 
+---
+# mazeGen 0.1.1
 
+### MINOR CHANGES
+
+* Change the wording in the maze instructions from 'top' to "{{direction}}" in the mazeHTML function and the mazeObject function. 
+
+* Updated the formula for mazeAbility function. 
+
+* Updated the formula for mazeDiff function to work properly. 
+
+* Updated an argument concerto so that users can create templates for both concerto version 4 and concerto version 5.